requirements.txt
Flask==3.1.0
flask-cors==5.0.1
tensorflow==2.19.0
numpy==1.26.4
pillow==11.2.1
scikit-learn==1.3.2
joblib==1.4.2
gunicorn
psutil
pandas